Get file content from Uri
