+++ app/src/main/res/values-de/strings.xml
@@ -68,5 +68,4 @@
     <string name="settings_info_libraries_explanation">Diese App nutzt Open-Source Bibliotheken, hier ist eine Liste aller genutzten Bibliotheken und deren jeweilige Lizenzen.</string>
 
     <string name="ini4j_license">Lizenz der [ini4j] Bibliothek (zum Ändern von Cuberite Einstellungen)</string>
-    <string name="pathutils_license">Lizenz der PathUtils Funktionen (genutzt um Dateien für die manuelle Installation zu wählen)</string>
 </resources>
