+++ app/src/main/res/values/strings.xml
@@ -102,22 +102,4 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
     </string>
-
-    <string name="pathutils_license">License for the PathUtils functions (used for installing Cuberite manually to select binaries)</string>
-    <string name="pathutils_license_description" translatable="false">A few edits were made to wrap some functions into a new class. Below is a copy of the original license.\n\n
-
-Copyright © 2011 - 2013 Paul Burke\n\n
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at\n\n
-
-  http://www.apache.org/licenses/LICENSE-2.0\n\n
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-    </string>
 </resources>
